Add depth limit as argument for how often to autotile
Currently the script will autotile (vsplit/hsplit) every window focused
on with no limit.

At a certain depth this behaviour is not desirably to me and I want to
add more windows to the same container. Especially for terminals I
prefer to have more space horizontally than vertically to see the entire
line if possible.

To achieve this, I add a optional parameter to prevent further
autotiling after the specified limit. For example with a 16:9 display
and a depth limit of 1, it will split horizontally first and then only
split vertically but in the same container.

```
 ________________
|        |       |
|--------|       |
|        |-------|
|--------|       |
|________|_______|
```

As a side effect, this enables stacking and tabbed layouts after
reaching the limit.
The default behavior is still the same with an infinite limit for how
deep to autotile.