Ignore containers for depth if they have 1 child node
With lots of splits the hierarchy can get deeper than normal, if
containers get closed. i3 does not flatten this hierarchy (atm).

This means that the depth option does needs to ignore containers that
only contain one child/node. This is done only incrementing the current
depth when checking against the limit if the nodes of the parent
container is bigger than 1.