Add information about --limit
The help and README did not go into details what exactly the --limit
option does.

Since it can be used to imitate a master-stack layout, and can enable
tabbed or stacking layouts to some degree, this adds the explanation in
the section talking about this limitation.

It softens the limitations of autotiling while still stressing that
tabbed and stacked layouts may behave unexpectedly and are not supported
by autotiling.